fix: main페이지 아티클 구조 변경
