+++ src/app/main/page.tsx
@@ -1,18 +1,25 @@
 import type { pageProps } from '@/types/page';
 import TodayTab from '@/app/main/TodayTab';
 import CategoryTab from '@/app/main/CategoryTab';
+import ArticleContent from './article/[id]/ArticleContent';
+import { GET } from '@/network';
+import type { ArticleType } from '@/types';
 
 const MainPage = async ({ searchParams }: pageProps) => {
-  const currentTab = (searchParams.tab ?? 'today') as string;
   // apiData에서 tab에 들어갈 데이터 fetch
+  const currentTab = (searchParams.tab ?? 'today') as string;
+  const articleApiData = await getMainPageArticleData();
 
   return (
     <div className='flex flex-col items-center w-full gap-10 mb-10'>
       {currentTab === 'today' ? (
         <>
-          <TodayTab />
-          <div id='article1' className='w-full h-[1200px] bg-darkgrey'></div>
-          <div id='article2' className='w-full h-[1200px] bg-blue'></div>
+          <TodayTab articleData={articleApiData} />
+          <div className='flex flex-col w-full gap-20'>
+            {articleApiData.map(article => (
+              <ArticleContent isToday={true} articleId={article.id} />
+            ))}
+          </div>
         </>
       ) : (
         <CategoryTab currentTab={currentTab} />
@@ -22,3 +29,8 @@ const MainPage = async ({ searchParams }: pageProps) => {
 };
 
 export default MainPage;
+
+const getMainPageArticleData = async (): Promise<ArticleType[]> => {
+  const response = await GET('/articleList');
+  return response.data;
+};
